+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org)
 
 ### Fixed
 * A possible segmentation fault when resizing the terminal.
+* Screens/item_details: correctly account for fields spanning multiple lines.
 
 ### [v0.7.0] - 2019-02-28